+++ src/mono/mono/metadata/image.c
@@ -18,15 +18,15 @@
 #include <errno.h>
 #include <time.h>
 #include <string.h>
-#include "image.h"
+#include <mono/metadata/image.h>
 #include "cil-coff.h"
 #include "mono-endian.h"
 #include "tabledefs.h"
-#include "tokentype.h"
+#include <mono/metadata/tokentype.h>
 #include "metadata-internals.h"
 #include "metadata-update.h"
 #include "profiler-private.h"
-#include "loader.h"
+#include <mono/metadata/loader.h>
 #include "marshal.h"
 #include "coree.h"
 #include <mono/metadata/exception-internals.h>
