+++ src/mono/mono/metadata/metadata.c
@@ -16,18 +16,18 @@
 #include <stdlib.h>
 #include <string.h>
 #include <glib.h>
-#include "metadata.h"
+#include <mono/metadata/metadata.h>
 #include "tabledefs.h"
 #include "mono-endian.h"
 #include "cil-coff.h"
-#include "tokentype.h"
+#include <mono/metadata/tokentype.h>
 #include "class-internals.h"
 #include "metadata-internals.h"
 #include "reflection-internals.h"
 #include "metadata-update.h"
-#include "class.h"
+#include <mono/metadata/class.h>
 #include "marshal.h"
-#include "debug-helpers.h"
+#include <mono/metadata/debug-helpers.h>
 #include "abi-details.h"
 #include "cominterop.h"
 #include "components.h"
