+++ src/mono/mono/metadata/sgen-mono-ilgen.c
@@ -21,7 +21,7 @@
 #include "metadata/marshal.h"
 #include "metadata/abi-details.h"
 #include "metadata/class-abi-details.h"
-#include "metadata/mono-gc.h"
+#include <mono/metadata/mono-gc.h>
 #include "metadata/runtime.h"
 #include "metadata/sgen-bridge-internals.h"
 #include "metadata/sgen-mono.h"
