+++ src/native/include/mono/metadata/appdomain.h
