+++ src/native/include/mono/metadata/assembly.h
