+++ src/native/include/mono/metadata/class.h
