+++ src/native/include/mono/metadata/debug-mono-symfile.h
