+++ src/native/include/mono/metadata/environment.h
