+++ src/native/include/mono/metadata/image.h
