+++ src/native/include/mono/metadata/loader.h
