+++ src/native/include/mono/metadata/mono-debug.h
