+++ src/native/include/mono/metadata/mono-gc.h
