+++ src/native/include/mono/metadata/mono-private-unstable.h
