+++ src/native/include/mono/metadata/object-forward.h
