+++ src/native/include/mono/metadata/profiler.h
