+++ src/native/include/mono/metadata/sgen-bridge.h
